+++ cli/cli/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `beam docs` creates and uploads CLI documentation to Readme.com
+- Added `project add` command that allows to create new project and add it to an existing solution
+
+### Changed
+
+- fix issues with failing `project new` command
+
 ## [1.16.0]
 
 ### Added